modularize normalizeLineEndings and detectCrLf
