+++ generators/bootstrap/utils.spec.mts
@@ -1,0 +1,46 @@
+import { jestExpect as expect } from 'mocha-expect-snapshot';
+import { writeFileSync } from 'fs';
+import { prepareTempDir } from '../../test/utils/utils.mjs';
+
+import { detectCrLf } from './utils.mjs';
+
+describe('bootstrap Utils', () => {
+  describe('::detectCrLf', () => {
+    describe('passing a crlf file', () => {
+      let cleanup;
+      before(() => {
+        cleanup = prepareTempDir();
+        writeFileSync('crlf.txt', 'a\r\ncrlf file');
+      });
+      after(() => cleanup());
+
+      it('should return true', async () => {
+        expect(await detectCrLf('crlf.txt')).toBe(true);
+      });
+    });
+    describe('passing a lf file', () => {
+      let cleanup;
+      before(() => {
+        cleanup = prepareTempDir();
+        writeFileSync('lf.txt', 'a\nlf file');
+      });
+      after(() => cleanup());
+
+      it('should return false', async () => {
+        expect(await detectCrLf('lf.txt')).toBe(false);
+      });
+    });
+    describe('passing a single line file', () => {
+      let cleanup;
+      before(() => {
+        cleanup = prepareTempDir();
+        writeFileSync('lf.txt', 'a single line file');
+      });
+      after(() => cleanup());
+
+      it('should return undefined', async () => {
+        expect(await detectCrLf('lf.txt')).toBe(undefined);
+      });
+    });
+  });
+});
